pbmanager: Move pbmanager to worker/protocols/manager/pb.py
