+++ master/buildbot/test/integration/test_worker_comm.py
@@ -27,7 +27,6 @@ from twisted.trial import unittest
 
 import buildbot
 from buildbot import config
-from buildbot import pbmanager
 from buildbot import worker
 from buildbot.process import botmaster
 from buildbot.process import builder
@@ -36,6 +35,7 @@ from buildbot.test.fake import fakemaster
 from buildbot.test.util.misc import TestReactorMixin
 from buildbot.util.eventual import eventually
 from buildbot.worker import manager as workermanager
+from buildbot.worker.protocols.manager.pb import PBManager
 
 PKI_DIR = util.sibpath(__file__, 'pki')
 
@@ -169,7 +169,7 @@ class TestWorkerComm(unittest.TestCase, TestReactorMixin):
 
         # set the worker port to a loopback address with unspecified
         # port
-        self.pbmanager = self.master.pbmanager = pbmanager.PBManager()
+        self.pbmanager = self.master.pbmanager = PBManager()
         yield self.pbmanager.setServiceParent(self.master)
 
         # remove the fakeServiceParent from fake service hierarchy, and replace
