+++ master/buildbot/test/integration/test_worker_workerside.py
@@ -25,7 +25,6 @@ from twisted.trial import unittest
 
 import buildbot_worker.bot
 from buildbot import config
-from buildbot import pbmanager
 from buildbot import worker
 from buildbot.process import botmaster
 from buildbot.process import builder
@@ -33,6 +32,7 @@ from buildbot.process import factory
 from buildbot.test.fake import fakemaster
 from buildbot.test.util.misc import TestReactorMixin
 from buildbot.worker import manager as workermanager
+from buildbot.worker.protocols.manager.pb import PBManager
 
 PKI_DIR = util.sibpath(__file__, 'pki')
 
@@ -117,7 +117,7 @@ class TestWorkerConnection(unittest.TestCase, TestReactorMixin):
                                              wantDb=True)
         # set the worker port to a loopback address with unspecified
         # port
-        self.pbmanager = self.master.pbmanager = pbmanager.PBManager()
+        self.pbmanager = self.master.pbmanager = PBManager()
         yield self.pbmanager.setServiceParent(self.master)
 
         # remove the fakeServiceParent from fake service hierarchy, and replace
